requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
APScheduler==3.10.4
reportlab==4.0.7
arabic-reshaper==3.0.0
python-bidi==0.4.2
openpyxl==3.1.2
pulp==2.7.0
selenium==4.15.2
webdriver-manager==4.0.1